add correct main.lua
